feature: Экран вопроса

+++ app/src/main/java/com/akimov/wordsfactory/screens/training/presentation/TrainingViewModel.kt
@@ -41,7 +41,7 @@ class TrainingViewModel(
                         wordsCount = (currentState as TrainingScreenState.EnoughWords).wordsCount,
                         trainingState =
                         TrainingState.Started(
-                            secondsToStart = second,
+                            secondsToStart = second as Int,
                         ),
                     )
                 }
